fix compilation broken by renames
