+++ packages/table/src/tableBody.tsx
@@ -22,7 +22,7 @@ import { ICellCoordinates } from "./common/cell";
 import * as Classes from "./common/classes";
 import { ContextMenuTargetWrapper } from "./common/contextMenuTargetWrapper";
 import { RenderMode } from "./common/renderMode";
-import { ICoordinateData } from "./interactions/draggable";
+import { ICoordinateData } from "./interactions/dragTypes";
 import { IContextMenuRenderer, MenuContext } from "./interactions/menus";
 import { DragSelectable, ISelectableProps } from "./interactions/selectable";
 import { ILocator } from "./locator";
